pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "shopee-async-api"
version = "0.2.0"
authors = [
  { name="ils15", email="support@ils15.com" },
]
description = "A modern, asynchronous Python wrapper for the Shopee API"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
]
dependencies = [
    "httpx>=0.23.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "anyio[trio]>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["shopee_async_api*"]

[tool.pytest.ini_options]
asyncio_mode = "strict"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
